Controllers and services updated

+++ NIR/src/main/java/ru/Mak/nir/controllers/UserController.java
@@ -1,5 +1,8 @@
 package ru.Mak.nir.controllers;
 
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import ru.Mak.nir.entities.User;
 import ru.Mak.nir.exceptions.UserAlreadyExistsException;
 import ru.Mak.nir.services.UserService;
@@ -14,34 +17,49 @@ public class UserController {
     @Autowired
     private UserService userService;
 
-    @GetMapping
-    public ResponseEntity userById(@RequestBody Long id) {
-        try {
-            return ResponseEntity.ok(userService.findById(id));
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body("User not found");
-        }
+    @RequestMapping(value = "{id}",
+            method = RequestMethod.GET,
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<User> userById(@RequestBody Long id) {
+        if (id == null)
+            return new ResponseEntity<User>(HttpStatus.BAD_REQUEST);
+
+        User user = userService.getById(id);
+
+        if (user == null)
+            return new ResponseEntity<User>(HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<User>(user, HttpStatus.OK);
     }
 
-    @PostMapping(consumes="application/json")
-    public ResponseEntity registration(@RequestBody User user) {
+    @RequestMapping(value = "",
+            method = RequestMethod.POST,
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<User> createUser(@RequestBody User user) {
+        if (user == null)
+            return new ResponseEntity<User>(HttpStatus.BAD_REQUEST);
+
         try {
-            userService.addUser(user);
-            return ResponseEntity.ok("Registration completed");
+            user = userService.addUser(user);
         } catch (UserAlreadyExistsException e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body("Registration error");
+            return new ResponseEntity<User>(HttpStatus.CONFLICT);
         }
+        return new ResponseEntity<User>(user, new HttpHeaders(), HttpStatus.CREATED);
     }
 
-    @DeleteMapping
-    public ResponseEntity deleteUser(@RequestBody Long id) {
-        try {
-            userService.delete(id);
-            return ResponseEntity.ok("User deleted");
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body("Registration error");
-        }
+    @RequestMapping(value = "{id}",
+            method = RequestMethod.DELETE,
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<User> deleteGoal(@PathVariable Long id) {
+        if (id == null)
+            return new ResponseEntity<User>(HttpStatus.BAD_REQUEST);
+
+        User user = userService.getById(id);
+
+        if (user == null)
+            return new ResponseEntity<User>(HttpStatus.NOT_FOUND);
+
+        userService.delete(id);
+        return new ResponseEntity<User>(HttpStatus.NO_CONTENT);
     }
 }
